Add BetweennessPartitioner, improved testing pipeline.
100% of new lines are covered, merging.

+++ tests/conftest.py
@@ -1,10 +1,8 @@
 """Module for test fixtures available for all test files"""
-from ast import literal_eval
-from configparser import ConfigParser
 from copy import deepcopy
 from functools import wraps
 from os import listdir, remove
-from os.path import getsize, join, exists, dirname
+from os.path import getsize, join, exists
 from shutil import rmtree
 
 import osmnx as ox
@@ -13,27 +11,32 @@ from networkx import set_node_attributes
 from requests.exceptions import ConnectTimeout
 from urllib3.exceptions import MaxRetryError
 
+from superblockify.config import (
+    TEST_DATA_PATH,
+    RESULTS_DIR,
+    PLACES_SMALL,
+    GRAPH_DIR,
+    HIDE_PLOTS,
+)
 from superblockify.partitioning import __all_partitioners__
 
-config = ConfigParser()
-config.read(join(dirname(__file__), "..", "config.ini"))
-TEST_DATA = config["tests"]["test_data_path"]
-RESULTS_DIR = config["general"]["results_dir"]
-
 ALL_CITIES_SORTED = sorted(
-    [city for city in listdir(f"{TEST_DATA}cities/") if city.endswith(".graphml")],
-    key=lambda city: getsize(f"{TEST_DATA}cities/" + city),
+    [city for city in listdir(f"{TEST_DATA_PATH}cities/") if city.endswith(".graphml")],
+    key=lambda city: getsize(f"{TEST_DATA_PATH}cities/" + city),
 )
 SMALL_CITIES = [
     city
-    for city in listdir(f"{TEST_DATA}cities/")
-    if city
-    in [city[0] + ".graphml" for city in literal_eval(config["tests"]["places_small"])]
+    for city in listdir(f"{TEST_DATA_PATH}cities/")
+    if city in [city[0] + ".graphml" for city in PLACES_SMALL]
 ]
 
 
 # Redefining names for extending fixtures
 # pylint: disable=redefined-outer-name
+
+# *****************************************************
+# Partitioner Fixtures
+# *****************************************************
 
 
 @pytest.fixture(
@@ -44,18 +47,31 @@ SMALL_CITIES = [
         else pytest.param(part, marks=pytest.mark.xfail(reason=part.__deprecated__))
         for part in __all_partitioners__
     ],
+    # use partitioner name, but cut off "Partitioner"-> "Part" if it is there
+    ids=[part.__name__.replace("Partitioner", "Part") for part in __all_partitioners__],
 )
 def partitioner_class(request):
     """Fixture for parametrizing all partitioners inheriting from BasePartitioner."""
     return request.param
 
 
-@pytest.fixture(scope="session", params=ALL_CITIES_SORTED)
+# *****************************************************
+# City Fixtures
+# *****************************************************
+# All cities
+# **********
+
+
+@pytest.fixture(
+    scope="session",
+    params=ALL_CITIES_SORTED,
+    ids=[city[:-8] for city in ALL_CITIES_SORTED],
+)
 def test_city_all(request):
     """Fixture for loading and parametrizing all city graphs from test_data."""
     # return request.param without .graphml
     return request.param[:-8], ox.load_graphml(
-        filepath=f"{TEST_DATA}cities/" + request.param
+        filepath=f"{TEST_DATA_PATH}cities/" + request.param
     )
 
 
@@ -66,25 +82,8 @@ def test_city_all_copy(test_city_all):
     return city_name, graph.copy()
 
 
-@pytest.fixture(scope="session", params=SMALL_CITIES)
-def test_city_small(request):
-    """Fixture for loading and parametrizing small city graphs from test_data."""
-    return request.param[:-8], ox.load_graphml(
-        filepath=f"{TEST_DATA}cities/" + request.param
-    )
-
-
-@pytest.fixture(scope="function")
-def test_city_small_copy(test_city_small):
-    """Fixture for getting a copy of small city graphs from test_data."""
-    city_name, graph = test_city_small
-    return city_name, graph.copy()
-
-
 @pytest.fixture(scope="session")
-def test_city_all_preloaded_save(
-    test_city_all, partitioner_class, _teardown_test_folders
-):
+def test_city_all_preloaded_save(test_city_all, partitioner_class):
     """Fixture for saving preloaded partitioners for all cities with bearing and
     length test_data. Without metrics. Shared across all tests."""
     city_name, graph = test_city_all
@@ -100,7 +99,7 @@ def test_city_all_preloaded_save(
 
 
 @pytest.fixture(scope="function")
-def test_city_all_preloaded(test_city_all_preloaded_save, _teardown_test_folders):
+def test_city_all_preloaded(test_city_all_preloaded_save):
     """Fixture for preloaded partitioners for all cities with bearing and length.
     Without metrics. Loaded for each test."""
     name, cls = test_city_all_preloaded_save
@@ -108,9 +107,7 @@ def test_city_all_preloaded(test_city_all_preloaded_save, _teardown_test_folders
 
 
 @pytest.fixture(scope="session")
-def test_city_all_precalculated_save(
-    test_city_all, partitioner_class, _teardown_test_folders
-):
+def test_city_all_precalculated_save(test_city_all, partitioner_class):
     """Fixture for saving precalculated partitioners for all cities with bearing and
     length test_data. Without metrics. Shared across all tests."""
     city_name, graph = test_city_all
@@ -132,10 +129,32 @@ def test_city_all_precalculated(test_city_all_precalculated_save):
     return cls.load(name=name)
 
 
+# ************
+# Small cities
+# ************
+
+
+@pytest.fixture(
+    scope="session", params=SMALL_CITIES, ids=[city[:-8] for city in SMALL_CITIES]
+)
+def test_city_small(request):
+    """Fixture for loading and parametrizing small city graphs from test_data."""
+    return request.param[:-8], ox.load_graphml(
+        filepath=f"{TEST_DATA_PATH}cities/" + request.param
+    )
+
+
+@pytest.fixture(scope="function")
+def test_city_small_copy(test_city_small):
+    """Fixture for getting a copy of small city graphs from test_data."""
+    city_name, graph = test_city_small
+    return city_name, graph.copy()
+
+
 @pytest.fixture(scope="session")
 def test_city_small_precalculated(test_city_small, partitioner_class):
     """Fixture for loading and parametrizing small cities with bearing and length
-    test_data. Without metrics."""
+    test_data. With metrics."""
     city_name, graph = test_city_small
     part = partitioner_class(
         name=f"{city_name}_{partitioner_class.__name__}_precalculated_test",
@@ -143,36 +162,16 @@ def test_city_small_precalculated(test_city_small, partitioner_class):
         graph=graph.copy(),
     )
     part.run(calculate_metrics=True)
-    return part
+    part.save(save_graph_copy=True)
+    return part.name, part.__class__
 
 
 @pytest.fixture(scope="function")
 def test_city_small_precalculated_copy(test_city_small_precalculated):
     """Return a copy of small cities with bearing and length test_data. Without
-    metrics."""
-    return deepcopy(test_city_small_precalculated)
-
-
-@pytest.fixture(scope="session")
-def test_one_city_precalculated(partitioner_class):
-    """Fixture for loading and parametrizing one small city with bearing and length
-    test_data. Without metrics."""
-    city_name, graph = SMALL_CITIES[0][:-8], ox.load_graphml(
-        filepath=f"{TEST_DATA}cities/" + SMALL_CITIES[0]
-    )
-    part = partitioner_class(
-        name=f"{city_name}_{partitioner_class.__name__}_precalculated_test",
-        city_name=city_name,
-        graph=graph.copy(),
-    )
-    part.run(calculate_metrics=False)
-    return part
-
-
-@pytest.fixture(scope="function")
-def test_one_city_precalculated_copy(test_one_city_precalculated):
-    """Return a copy of one city with bearing and length test_data. Without metrics."""
-    return deepcopy(test_one_city_precalculated)
+    metrics. Loaded for each test."""
+    name, cls = test_city_small_precalculated
+    return cls.load(name=name)
 
 
 @pytest.fixture(scope="session")
@@ -213,13 +212,82 @@ def test_city_small_osmid_copy(test_city_small_osmid):
     return test_city_small_osmid.copy()
 
 
+# ********
+# One city
+# ********
+
+
+@pytest.fixture(scope="session")
+def test_one_city():
+    """Fixture for loading and parametrizing one small city."""
+    return SMALL_CITIES[0][:-8], ox.load_graphml(
+        filepath=f"{TEST_DATA_PATH}cities/" + SMALL_CITIES[0]
+    )
+
+
+@pytest.fixture(scope="function")
+def test_one_city_copy(test_one_city):
+    """Fixture for getting a copy of one small city."""
+    city_name, graph = test_one_city
+    return city_name, graph.copy()
+
+
+@pytest.fixture(scope="session")
+def test_one_city_precalculated(partitioner_class, test_one_city):
+    """Fixture for loading and parametrizing one small city with bearing and length
+    test_data. Without metrics."""
+    city_name, graph = test_one_city
+    part = partitioner_class(
+        name=f"{city_name}_{partitioner_class.__name__}_precalculated_test",
+        city_name=city_name,
+        graph=graph.copy(),
+    )
+    part.run(calculate_metrics=False)
+    part.save(save_graph_copy=True)
+    return part.name, part.__class__
+
+
+@pytest.fixture(scope="function")
+def test_one_city_precalculated_copy(test_one_city_precalculated):
+    """Return a copy of one city with bearing and length test_data. Without metrics.
+    Loaded for each test."""
+    name, cls = test_one_city_precalculated
+    return cls.load(name=name)
+
+
+@pytest.fixture(scope="session")
+def test_one_city_preloaded(partitioner_class, test_one_city):
+    """Fixture for loading and parametrizing one small city not run yet."""
+    city_name, graph = test_one_city
+    part = partitioner_class(
+        name=f"{city_name}_{partitioner_class.__name__}_preloaded_test",
+        city_name=city_name,
+        graph=graph.copy(),
+    )
+    part.save(save_graph_copy=False)
+    return part
+
+
+@pytest.fixture(scope="function")
+def test_one_city_preloaded_copy(test_one_city_preloaded):
+    """Return a copy of one city not run yet."""
+    return deepcopy(test_one_city_preloaded)
+
+
+# *****************************************************
+# Other Fixtures
+# *****************************************************
+# Clean up test folders after tests are done
+# ******************************************
+
+
 @pytest.fixture(scope="class")
 def _teardown_test_graph_io():
     """Delete Adliswil_tmp.graphml file and directory."""
     yield None
     work_cities = ["Adliswil_tmp", "Adliswil_tmp_save_load"]
     for city in work_cities:
-        test_graph = join(config["general"]["graph_dir"], city + ".graphml")
+        test_graph = join(GRAPH_DIR, city + ".graphml")
         if exists(test_graph):
             remove(test_graph)
         results_dir = join(RESULTS_DIR, city + "_name")
@@ -232,17 +300,28 @@ def _teardown_test_folders():
     """Delete all test data folders."""
     yield None
     # Delete all folders in RESULTS_DIR that end with _test
-    for folder in listdir(RESULTS_DIR):
-        if folder.endswith("_test"):
-            rmtree(join(RESULTS_DIR, folder))
+    if exists(RESULTS_DIR):
+        for folder in listdir(RESULTS_DIR):
+            if folder.endswith("_test"):
+                rmtree(join(RESULTS_DIR, folder))
 
 
-@pytest.fixture(scope="function", autouse=config.getboolean("tests", "hide_plots"))
+# ***********************
+# Hide plots during tests
+# ***********************
+
+
+@pytest.fixture(scope="function", autouse=HIDE_PLOTS)
 def _patch_plt_show(monkeypatch):
     """Patch plt.show() and plt.Figure.show() to prevent plots from showing during
     tests."""
     monkeypatch.setattr("matplotlib.pyplot.show", lambda: None)
     monkeypatch.setattr("matplotlib.pyplot.Figure.show", lambda _: None)
+
+
+# ****************************
+# Mark download tests as xfail
+# ****************************
 
 
 def mark_xfail_flaky_download(test_func):
